# PERF: faster indexing for non-fastpath groupby ops